add a single  "/" (U+002F) character to void element
to be more cleann add the / to void element as described in https://www.w3.org/TR/html5/syntax.html#start-tags point 6.

Then, if the element is one of the void elements, or if the element is a foreign element, then there may be a single "/" (U+002F) character. This character has no effect on void elements, but on foreign elements it marks the start tag as self-closing.